+++ packages/cli/CHANGELOG.md
@@ -10,6 +10,28 @@
   - More accurately monitors signals (ctrl+c) and shutdowns.
   - Tasks can now be configured with a timeout.
 
+## Unreleased
+
+#### 🚀 Updates
+
+- Added `configuration` and `scaffolding` variants to the project `type` setting in `moon.yml`.
+  - Updated project constraints to support these new variants.
+- Added a `stack` setting to `moon.yml`, for categorizing which tech stack it belongs to.
+  - Supports `frontend`, `backend`, `infrastructure`, and `systems`.
+  - Added a `projectStack` field to the query language (MQL).
+  - Added a `$projectStack` token variable for tasks.
+  - Updated the `moon query projects` command to support a `--stack` option, and include the stack
+    in the output.
+  - Updated the `moon project` command to include the stack in the output.
+- Updated task input environment variables to support a wildcard match using `*`, for example
+  `$VITE_*`.
+  - This will include all environment variables in the current process that starts with `VITE_`.
+- Updated the `envFile` task option to support a list of file paths.
+
+#### ⚙️ Internal
+
+- Updated Rust to v1.76.
+
 ## 1.21.3
 
 #### 🐞 Fixes